tests/lint: clang -Wshorten-64-to-32 narrowing lint + baseline (#4a)
The LP64 bug class is a 64-bit value silently narrowed to 32 bits crossing a
boundary; clang's -Wshorten-64-to-32 is exactly that warning and gcc has no
equivalent. Add a clang analysis pass (gcc stays the production compiler):
tests/lint/run.sh asks `mk -n -a` for the real per-file compile flags of every
host C file (libs + emu), replays each through clang -fsyntax-only with only
that warning enabled, and diffs against a curated baseline so NEW narrowings
stand out from the 246 pre-existing (mostly benign) ones.

make lint / lint-all / lint-update. Baseline keyed by file+conversion-kind so
it doesn't churn on line moves. Also makes aw() in os.c clean (write returns
ssize_t -> use long), so the new observability code adds no narrowings.

+++ emu/Linux/os.c
@@ -73,7 +73,7 @@ static char *disstate[] = {	/* must track enum ProgState in interp.h */
 static void
 aw(char *s, int n)
 {
-	int r;
+	long r;
 
 	while(n > 0){
 		r = write(2, s, n);
